packages/server/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.9.2
blinker==1.9.0
certifi==2025.8.3
click==8.3.0
distro==1.9.0
fastapi==0.118.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.0
MarkupSafe==3.0.3
openai==2.0.1
pydantic==2.11.9
pydantic_core==2.33.2
python-dotenv==1.1.1
sniffio==1.3.1
starlette==0.48.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.37.0
Werkzeug==3.1.3
